Adds the tests and implementation of applySingleCommand method
Adds the tests for the applySingleCommand method which decides
if the command is turn or move command. The applyMoveCommand and
applyTurnCommands return types are changed which were directly
updating the values for testing purposes. Now they return the coordinates
or the direction, which are used to update the rovers values in the
applySingleCommand method.

Implements-Requirement: Enter requirement-id or delete line
Signed-off-by: emreoezmen <emreykal@gmail.com>